requirements.txt
django
djangorestframework
markdown
django-filter
drf_yasg
requests